Add automatic termination logic
The counter has been repurposed to count how many
non-finished tasks there are. The current solution
can correctly identify when to finish execution.

Update termination-logic to not crash scheduler

By moving the termination logic the scheduler will
properly finish without crashing the program.

Allow partial schedule

Previously the user had to define the entire schedule
that was supposed to be followed, from start to finish.
However, this commits makes it so that a starting schedule
may be chosen and then it leaves the remaining choices
to the scheduler.

Move termination signal to scheduler thread

Previously the termination signal was sent from whatever
task that was the last to execute. This was not ideal,
since it could result in the scheduler thread to be stuck,
even after termination. This is fixed by moving the
termination signal to the scheduler, and the last task
instead signals the scheduler that something has happened to
the queue.

Change if-else into if with return

The previous solution to completing the loop and sending the
termination signal was by having an if-else. If the termination
signal is sent, then no other part of the loop is executed.
This was to prevent the scheduler from crashing because it was
trying to access an non-existing task. This was added unnecessary
complexity and indentation. The scheduler task can instead
directly return after sending the termination signal.