Add unique ID to each task
This commit now adds an ID variable to each task, which keeps track of tasks ID.
To get this to work, the ID also keeps track of the number of children the task has, since otherwise it would be hard to get the correct format.

Add functionality for final nested future .0.

The flag `isEnd` can be set to true, to set the ID of the task to `parent_id.0.`