Record scheduler decisions
Adds a simple way for the scheduler to record which
decisions it has taken. Right now this should suffice,
however in the future we may be interested in the
previous runs (if we run the scheduler multiple times)?
In that case this code needs to be updated.

Automatically schedule final child

As discussed in the previous meeting adding a
"final" child with identifier `parent_id.0` can
help us when replaying a schedule since it will
allow us to make sure that a task (and all of its subtasks)
finish before continuing. This commits makes it so that
this "final" task child is automatically added right
before the task is completed. The current implementation is
safe in the way that the "final" task child will never
execute before the task or its other children has finished.