Replay scheduler decisions
Adds the ability for the scheduler to follow a
predetermined schedule. Right now the solution makes
use of a list. There could be more efficient ways,
sine the lists are singly-linked. A better approach
may be to use some sort of tree, in which the `id`
will tell which branch to take. Since we know that a
parents id will always be a substring of the child id.