Add ability to wait for maximum number of tasks
The scheduler now waits until it knows that it must
make a choice, since it is impossible to continue otherwise.
This is done by adding one `AtomicBoolean` and one condition.
The `AtomicBoolean` is necessary to signal if we should execute
the scheduler as normal. This can happen in two cases, either we
have gotten to the end (`awaitTermination`) or we call `await`
on a top level task. The reason why it is not enough to simply
send a signal is that that would only allow a singular choice.
However, there is no guarantee that we would make the "correct"
choice of `awaited` task.

Fix bug with main thread stalling after termination

There was a bug in which if we `awaited` a top level
task, but had other tasks ready then we may call `termination`
before the main thread has properly started. This can
result in the scheduler thread being signaled and tricking
it into terminating. While the main thread is still waiting.
This could be solved by making sure that `cnt` will never be
zero while we wait for a `top level` task.

Add comments to explain changes and functionality

Add tests for waiting until queue is filled

To increase the confidence in the logic for making
the scheduler wait as long as possible to make a
choice, more unit tests have been added.