Add pluggable algorithms
Adds a trait called `ExplorationAlgorithm` which has
two defined methods. `getNext()` and `prepareNext()`.
`getNext()` takes the readyQueue and returns an option,
of either the next task and its controller or none.
None is chosen if the algorithm decides there is no
"good" choice.

Together with this two algorithms have been implemented.
One is a random walk algorithm, which only makes a random choice.
The second algorithm is the `FixedSchedule` algorithm,
which has been defined/used before. However it has been
adapted to use the trait.

Add random walk algorithm

Add an algorithm which at each decision point chooses
a random ready task to continue.

Add Fixed Schedule algorithm

The fixed schedule algorithm which previously was
a part of the Scheduler object has now been moved
to be its own class that implements the `ExplorationAlgorithm`
trait. The reasoning behind it being a class is that
we may want multiple different fixed schedules.

Furthermore, moving it into its own class/object makes
the code easier to maintain and understand. The biggest
change needed for this is to change the return type
of the `getNext` to an option such that if the
`ExplorationAlgorithm` is not satisfied with the current
`readyTask` queue it may signal the scheduler to wait
for a change in the queue.

Add tests for RandomWalk algorithm

To increase the confidence that the RandomWalk
algorithm works as intended, some test cases have
been added which uses this algorithm.