Add github actions
Adds an YML file which will enable github actions to
test the code after each pull-request to the main branch.

Currently, the tests are only run on the latest version
of ubuntu. This should be enough especially since the
code should work the same on all operating systems.
However, it should be very easy to add multiple
operating systems if necessary.

Fix bug where scheduler could stall indefinitely

Since the lock is unlocked at the end of each loop
before being reacquired in the start of each loop there
is a slight timeframe in which multiple tasks could finish.
If this would lead to the remaining total tasks to become
zero, then the scheduler would stall since it would
arrive at `awaitUninterruptibly` without any possibility
of getting a signal.

This can be fixed by setting `done` to true in the
finish function if we should finish, since that would
ensure that the `awaitUninterruptibly` is never called.