Add stress tests
To provide better confidence in the logic stress tests
have been added. This is both to monitor how the scheduler
performs under a heavier load, and to see that repetetive
use of the scheduler does not affect the final result.

The added tests are in two categories.

1. Heavy load
2. Repetetive use

For the heavy load tests, a number (1000) futures are
created and then the tests make sure that the execution
follows as it should.

For the repetetive use tests, the schedulers previous tests
are repetetivly called 1000 times each (for some select
tests).

Fix faulty termination bug

Previously, there was a bug in which one run of the
scheduler could affect the next run of the scheduler.
This bug occured because of faulty termination logic,
which made the scheduler reset before it had a chance
to terminate. This caused the previous scheduler task
to continue running, which could cause problems.

This was fixed by remaking and restructuring the termination
logic, to make sure that all termination is safe.
Furthermore, the scheduler may now only terminate after
the `awaitTermination` function has been called.