Add scalafmt
Add scalafmt and a check to the github actions that
make sure that the project is formatted according to
the config file, or an error will be thrown.

This means that it is the users responsibility to format
the code before a push/pull request. This can be done
with the `sbt scalafmtAll` command.

The check is then tested with `sbt scalafmtCheck`

Format scala files

Format scala files according to the config file