Add functionality for file I/O and error handling
In order for a user to get use out of a previous schedule
it is helpful if that schedule was in some form persistent.
Previously, this was not the case since it was just saved
as a variable. Now functions have been added to write
a schedule to file, and to read a schedule from file.

In conjunction with this the ability for tasks and the
scheduler to handle errors and exceptions have also been
added. A task that would throw an exception will increase
signal the scheduler that an exception was thrown and
terminate. Then if by user input the scheduler can
check if any exceptions have been encountered, and if
they have been the schedule will be written to file.