Add base for code coverage tool
Added some tools to help users get the code coverage
of a test (or code) when using the scheduler.
This consists of adding a `CoverageTracker` object
which has a map which updates a specific ID each time
the `marker` function is called. This allows the user
to then know which markers were triggered and how often.

Add coverage tracker tests

To be sure that the coverage tracker works as intended
some tests have been added that should work for all
possible exploration algorithms (given that they are valid)

Add ability for user to add expected markers

To be able to know which markers were missed the
`coverageTracker` must be able to know which markers
to expect. Right now this is the job of the user to
provide, but may be changed in the future.