Add `checkSuspend` and sequential execution
Add `checkSuspend` function

To catch and accurately model some errors, it is necessary
to have fine-grained control over in what order concurrent
operations are taken. The `checkSuspend` function allows
this by suspending the task, and adding it back to the
list of ready tasks. This allows the scheduler to explore
many kinds of interleavings with greater control.

Add sequential execution

Adds the ability for the scheduler to run tasks sequentially.
This is done by only allowing one future to progress at
a time. Which then makes the code behave as the futures
were executed sequentially.

To give more control the `checkSuspend` function can
be used to suspend a future, possibly allowing another
future to continue instead.