Add function to check reliability of schedule
Some schedules may not be reliable in the sense that
there may be some non-determinism which determines if
that schedule actually leads to problematic behaviour.

This means that the problematic behaviour that was
discovered may only appear rarely in that schedule,
and therefore that schedule is not an ideal candidate
to replay the problem.

The added function replays the schedule a number of times
checks if it gets any errors and that the output is
as it expects. Then if the correct behaviour is seen
for a certain percentage of all runs, then we know that
the schedule is reliable.