Add barber shop test and schedule controlled locks
Add barber shop test

Adds a really simple barber problem example in the form
of a test. The test uses one `badcustomer`. The `badcustomer`
has a faulty lock implementation, which can lead to race-
conditions. The test then checks if the `CoverageTracker`
can find all markers, and then checks if the sequential
replay of the schedule is reliable.

The test also contains an `okcustomer` which is the
correct implementation of `badcustomer`.

Add custom scheduler lock and condition

Previously a users only way of controlling the execution
of a program with the scheduler was via `future.await`.
This can be helpful, but is in many cases not enough.
For example when multiple futures are accesing the same
resources. In these cases locks are needed, to make the
resources mutually exclusive. When running code in parallel
mode, this is not a problem since if a task does not have
access to a lock, it will simply wait until the task that
has the lock will release it.

However, in the sequential mode this is not enough since
only one task will be running at a time. This means that
if a task has the lock, but then has to stop execution
for some reason then all other tasks that want that lock
will not be able to progress. Which can cause the program
to stall.

This problem is solved by adding a custom lock and condition
that checks if the lock can be locked, and locks it. Otherwise,
it resubmits the task to the scheduler and tries again
when the scheduler schedules it again, until the lock is
free.