Add deadlock recognition with `checkSuspend`
Add the functionality for the scheduler to alert and
in most cases continue if it suspects that a deadlock
has occured. The deadlock check is done via the
`checkSuspend` function. The user can define a boolean
flag that describes if the `checkSuspend` should
start a new timeout timer or remove the existing one.
This allows the scheduler to signal after a certain
amount of time, which will interrupt all running tasks.

For the tasks that are in a deadlock this will in most
cases cause an `InterruptedException` which will cause
the tasks to terminate.

Right now this will just result in an error, as long as
the main thread does not need the information from the
future. This is gained by calling `await` on the future
(or a child future has timeout and the main thread waits
for the parent future), which will then throw an error.
In other cases were the future does not return anything,
and is not `awaited` then the error counter will be
incremented. However, the main thread will not throw
an error.