Add Controller class with `CyclicBarrier`
Previously, a new `promise` was created each time a task had
to wait for the scheduler to signal it. This works but is not
ideal, since a new one has to be constructed each time.
Furthermore, future development to add, for example, actor
support is hard with the current setup. Since most functions
require a task.

By creating a `controller` class, it is both possible to easily
use `CyclicBarriers` and to have all necessary (universal)
information in one place, like the id. This allows the
scheduler to reuse the same `CyclicBarrier` thereby saving
resources and makes future developments easier.